Separated poll sections into own components
